Validate the repair results, only use clean output
The bot repair routine doesn't always manage to produce an output that
satisfies the thin bots check, even when it produces a manifold output.
Check the repaired output, and only replace the original manifold result
with the new BoT when it passes the thin faces check.  Note in the
verbose log what happened.

We need to be more aggressive in the librt routine - filtering out
first-hit misses results in our not seeing some problems that the repair
routine can handle (the havoc windows are a case in point.)