Add previous facetize logic back in as facetize_old
Because the new facetize system is such a radical departure from older
codes, add the 7.38.2 facetize logic to have as a backup under the
command name facetize_old.  This can go away as we get more confidence
in the new setup, so it's not being announced as a command per-say, but
it will be available to try if the new logic isn't working for some
reason.