Take the perturb routines for a test drive with qged.
Not a cure-all, apparently, but potentially useful none the less.
Working with r.ant5 from havoc, perturbing the primitives and then
facetizing produces a bot that still won't pass lint, but seems to have
fewer failures.  Following up with bot strip, bot repair and bot
decimate was able to get to a clean BoT.