fix flat search + exec bug
when doing a flat search with | in mged, the path was being built with
db_path_to_string which always assumes the path is starting from the
root - and therefore prepends a forward slash. However, this is not
always the case. Bug was noticed when using a 'search | .. -exec ..' as
the exec inputs would have a leading slash and would break.