naive fix for bot raytracing on Windows
with the recent hlbvh work merge, Windows raytracing for bots was
broken. This was caused by indices being based on the current thread id
(gathered with bu_thread_id()). On Windows, the thread ID is often
recycled from recently closed threads which means indexing into the
hit_array is not guaranteed unique. This lead to bad rt images and
memory access vioations.

To temporarily fix the issue, we need to allocate the hit_da container
on the stack before doing shot calculations. This will almost certainly
have some form of performance implication but the extent has not yet
been tested.