fix bu_file test failure by adapting our test.
test previously assumed we're running on a unix-like filesystem with distinct read/write/execute bits that are set a particular way by default.  this was both wrong for different umask settings as well as different filesystems (e.g., exfat) that are fixed-permissive (akin to immutable 777 everywhere).  thus, changed the test to force the files permissive with an bu_fchmod(777) where we can at least ensure read/write/execute are set (and we can't reliably check if unset after bu_fchmod(000) or variations thereof without knowing the FS).

also added a few more checks to make sure files exist and do not exist at the expected times.