refactor file+dir teardown into a function to ensure clean start.
encountered some build dirs were leaving files around, which can happen when the test fails or the testing is interrupted.  so this ensures we delete what we expect to create if it exists at the very start (as well as at the end when we pass).  that way files will still be there for inspection if/when the test fails (hence no atexit()).