Update libbg SPSR to use upstream header-only mode
Not working yet, but we seem to be able to feed our data in using this
approach and get *something* out, so checkpoint now to document how to
use their templates and streams.  Probably can refine this further, but
it does compile and run.

At the moment the decimation routine explodes with a bunch of SHOULD NOT
HAPPEN errors and crashes super hard (brings down the desktop, in fact.)