add simple progress bar feedback
when the execution got moved to gchecker (fdd3918), the progress bar for
this tool essentially became useless. In a perfect world we'd probably
have some sort of signal to poll progress and report back or refactor
gchecker to run one suite at a time. But, tcl doesn't easily accomodate
signal sending, and a refactor would be overkill just for a progress
bar.

In the meantime, give very basic status text to let the user know that
their click is doing *something*.