Test using upstream GeometricTools directly for lseg_lseg
Trying out using the C++ header-only GeometricTools implementation under
the hood of libbg.

We originally translated the code to C for direct embedding, but since
we're now using newer C++ standards in our own code we may be able to
replace all of our embedded copies of the GeometricTools algorithms with
direct usages (as well as leveraging a number of other algorithms
present in the GeometricTools code - in particular, their oriented
bounding box and convex hull implementations may be good candidates to
replace our current 3rd party codes.)

Deliberately starting small to get a sense of the implications, but
leveraging the upstream maintained code in favor of our refactored
copies is desirable - these sorts of problems can be tricky in corner
case scenarios, so it would be advantageous to leverage any upstream
improvements in the algorithms.