exit if no layout can be found
if we get to this point and something has gone wrong (ie it is unable to
find a layout), we need to abort.

Current problem was observed with our m35 model. The tops objects are
'scenes' and have the component sitting on a halfspace groundplane with
lights. The halfspace in particular screws things up as we rely on the
bounding box size (which is inf) and orthogonal views (bottom view is a
blank white square).
This is a tricky problem to solve auto-magically. So for now, just punt
and make the user manually specify the correct component.