Disable test exec builds if BUILD_TESTING=OFF
Allow for a more minimal compilation if testing is off - disable the
compilation of test executables based on CTest's BUILD_TESTING variable.
If that variable is off CTest tests won't be available in any case.
Allows for more minimalist compilations as well when using
BRLCAD_COMPONENTS, which is the original user case prompting the suggest
to support disabling them.