Don't apply the executable mask on Windows
Per Microsoft's _chmod documentation, their version doesn't have a
concept of executable, so "setting" it with bu_fchmod is a no-op.  We're
getting a path on Windows now, but the is-executable test was still
failing because the files sb.st_mode didn't match the mask we were
testing.

This may point to a problem with our config_win.h defines for this case,
and if we're willing to dive much deeper into the ACL permissions setup
there might be analogs we could map this to, but for now go the simple
route and just don't check for something we can't set with _chmod.