fix the 'x' aka 'solid_report' command.
the command is set up to take 'levels' which can be negative numbers, which indicate reporting of just leaf nodes or just combs, etc.  The assumption that a dash meant help meant any negative levels throw usage.  MGED uses x -1 and x -2 for the solid and object/comb editing panels (for object selection).