Correct memory managment for BRL-CAD brep object
ON_ModelGeometryComponent* CreateForExperts documentation in OpenNURBS
header opennurbs_model_geometry.h indicates that a true value in the
first argument indicates the geometry object should be deleted by
the ON_ModelGeometryComponent destructor.  That's not what we want -
we're managing the brep object ourselves - so we need this to be false.

This may be the reason doing a "sed" operation on a brep was crashing
MGED, but needs more testing to confirm this has actually fixed it.