Adjust bext text file processing for CMake find_package
Normally CMake's find_package *Config.cmake files don't hardcode paths,
but it looks like our z_brl library path made it in as a full path.  The
standard text processing set this path to the final install path, which
wasn't present, and that caused problems.  Moreover, the pkgconfig files
(which can be used by CMake, depending on how it searches) do use full
paths, and so would have similar issues.

Adjust the path setting logic so the CMake input files use the build
directory path while they are in the build directory, and add a
post-install rule to correct them for their final install location. We
could in principle do this for all text files, and that might be the
simpler route long term, but it would slow the install step
significantly to post-process ALL text files (particularly when large
dependencies like LLVM and Appleseed are added to the mix) so for now we
target this specifically at find_package inputs.