Remove intermediate mesh checks in bot extrude
Learned from upstream that these checks can be very expensive.  If this
type of debugging becomes necessary, introduce it when working the
problem rather than paying the overhead all the time.

(On a semi-related note, there's a bug in the upstream RotateUp routine
which appears to be behind the problems with using their Cylinder, so
we'll watch for developments there - would still be nice to use their
implementation (which is what their testing will also use) rather than
our custom one-off logic.)