fix a variety of crash cases working with old bspline nurbs.
this is just enough to get iges-g to succeed without crashing or bombing, but there does still appear to be some incorrect book-keeping going on.  there are paths through the logic were specific iges entities are ending up without geometry or without a mate. there were also cases resulting in loopuse without edgeuse, though now a minimum was put in place to progress through the existing logic.  this requires a much more careful study as iges-g -t now completes but doesn't result in a usable conversion.