Place setup logic inside main function
If I'm understanding CMake's rules for this, variables defined in
functions will remain defined when calling nested functions.  If that is
the case, we can just set the variables we need in our top level
function and avoid the "global to file" setup we had previously without
having to explicitly pass all the values we need as arguments to the
child functions.  That's only marginally better than a global, to be
honest, but at least the setup code is not unencapsulated logic that'll
be run on include()