Rework configure's handling of bext
Changing the logic to use src/bext as the source of the SHA1 used for
building, when BRL-CAD's configure process is responsible for producing
bext binaries.  This will allow checkouts of older BRL-CAD releases to
pull versions of bext that will work with their checkouts.  Previous
releases were pulling the RELEASE branch's current state, which may be
wildly out of sync if the cloned BRL-CAD version is older.

Also switching to using main as the touchpoint for BRL-CAD, rather than
the RELEASE branch.  For the bext repository itself, the plan is to
switch to a more typical github workflow of doing major updates in
branches and then merging them into main once complete, rather than
trying to keep main and RELEASE in sync.

Another feature added in this rework is the ability to recognize when a
configure-built bext is out of date.  The logic will now automatically
clear out the old build products and set the stage for a new build.
(Experience has shown us that if we don't do this automatically it's
quite common for the user to need to do it themselves due to bext
changes being necessary for newer BRL-CAD code to compile.)