Document a way to submodule update w/ local mirror
Since the URLs are baked into git submodule definitions in the history,
git doesn't know by default how to go looking for a local checkout.
Fortunately, it looks like the url insteadOf mechanism is able to
address this as long as the on-disk hierarchy of our mirror matches that
of the github URLs (which isn't hard, since it's all flat.) It's really
pretty simple - we just swap out the github prefix of the URL with our
local path and change the protocol.

When the submodule isn't populated yet we appear to have two options -
pass the necessary settings in via the command line -c options, or set
them in the global config file.  (Since the submodules aren't
initialized yet we can't use their local .git/config files.) If we don't
want to supply the -c options it  may be worth checking into the new
includeIf mechanism to see if it could help localize the impact of such
settings: https://stackoverflow.com/a/71096731

As long as github.com is viable this won't be an issue, but if we ever
have to change hosting services or work strictly with a local mirror the
submodule URLs *will* be a problem.  We can't change them in our history
without rewriting it in a major (and sophisticated) way - we would first
have to rewrite all of the bext history to use new URLs and update all
SHA1 references in submodule definitions, comments and content while
maintaining a map of old commits to new.  Once we had that map we would
need to do the same thing with the main brlcad history.  All things
considered, remapping the URLs seems to be far easier, less invasive and
effective.