Add more detailed instructions for URL remapping with git
If we ever end up changing hosting solutions, having to work with local copies
of the github.com repositories, etc. these notes will be needed.

This is the alternative to doing a gnarly history rewrite to update submodule
URLs - as long as a basic prefix subsitution on the URLs is sufficient, this
appears to do the job.