Fix Archer Center View button (vslew GED cmd)
Got a user report that the Archer Center View button was broken as of
BRL-CAD release 7.34.2.  Looking into it, the vslew command it is using
was looking for a GED command that wasn't known by that name.  We want
to call the libged slew plugin with that command, so we just have the
plugin define "vslew" as one of the acceptable strings to run it.

Looks like this probably got broken with 5f8aa3204 changing the ged_slew
call out for a ged_exec.  In this particular case the command argv[0]
didn't end up corresponding to the ged_* function signature it was
replacing, since the libtclcad command was "vslew" rather than "slew".

Looking through the Tcl code for $mGed commands, it looks like there may
be others that don't have active libged definitions.  Need to
investigate.  This is a textbook illustration of why interpreted
languages are fragile for this sort of work...