Initial test of automatic functions for ged cmds
This tests out a mechanism to set up automatic compile-time sensitivity
to situations where commands are renamed or removed, but are being
execed by other LIBGED commands.

We have a scanner read the plugin source code files to determine the
active GED command names.  Using that list, we then define a header and
wrapper functions.  The wrapper functions can be called in lieu of a
direct ged_exec.  When the code is changed and BRL-CAD is next
reconfigured, the newly generated function set will reflect the command
changes and callers of the wrappers will see a compile failure.

This isn't a full conversion of libged to the new approach, just a few
commands to test it out.  The goal here is to avoid a situation were we
don't find out seldom-exercised commands are broken until many releases
later when someone tries to run it.  It's not perfect, obviously, in
that changed command options can still break things, but at least this
way we will have some minimal detection of changes which are guaranteed
to break things.