Make argv[0] string setting optional
If the ged_exec_* function name encodes the command name, there's no
reason to force the user code to set it.  There's no particular harm if
they need or want to for a specific command, so allow it, but we have
the info to (literally) fill in the blank so go ahead and do so if there
is one.

The advantage in doing this (beyond cleaner client code) is that a
command defined in this fashion can be automatically renamed
successfully with find -exec and perl.  For example, if rm subsumes
killtree then s/ged_exec_killtree/ged_exec_rm is a one-and-done if the
calling code isn't also setting av[0] = "killtree";