Adjust how the ged_cmds are included.
Simplify the prefix, use BRLCADBUILD to guard a more general inclusion
of ged_cmds.h rather than doing it per file, and use a template for
ged_cmds.h so we can provide better explanation.

Placing ged_cmds.h in a general include location so it can potentially
be used in other BRL-CAD code - anywhere we are hard-coding a ged_exec
of a specific command we really should be doing this to enlist the
compiler's help in catching renaming/removal errors.  Whether it should
be actual public API is a separate concern, but as long as we don't
install it and guard its include with BRLCADBUILD it's internal only.