Convert the readily convertible cases of ged_exec to wrappers
For any cases in our code where we are using ged_exec simply to fire off
a pre-defined command, use the wrapper so we will get a compile time
breakage in the case of a command rename or removal.

Cases where we are still using ged_exec warrant scrutiny to make sure
it's a correct usage.  There are some (gsh and qged command lines, for
example) where the intent *is* general command execution, and a couple
cases where there are multiple potential command names coming in from
the caller that are passed through as-is (though the latter are
problematic in that we're still losing the compile time validation.)
There are definitely a few cases in libtclcad that are straight up
broken - I tried to label them as I encountered them.