Add in compat code to make db_search change minimally impacting, deprecate old signature
This is a breaking change if we just yank the old db_search_context type and
change the db_search signature, so we'll put in the older code under the
db_search_old name, list it in CHANGES, and do the deprecation for the previous
signature.  This lets us have our code use the new signature while still
allowing a regex to keep the old code working for a while.