List libged "how" command for deprecation, add replacment.
Yuck yuck yuck.  In at least some cases this appears to be returning an
internal private libged numerical #define value as textual output.
Should be human understandable, not using a number, and I don't think
this deserves to be a top level GED command anyway.  Plan is to use a
view obj [name] info mode subcommand to report this instead.