Consolidate platform specific fbserv functions
Trying to keep everything in sync during the refactor was getting
annoying, so use a Tcl Channel based define and consolidate the two
different sets of functions, using ifdef USE_TCL_CHAN to select the
necessary logic just in those cases where it actually differs.

While we're at it, don't hide the dm_netfd and dm_netchan globals behind
defines - we'll need to deal with those at some point, so make it
clearer what they are.  As a bonus that clears a WIN32 ifdef.