Make GEDP a non-global variable in MGED
The global state in MGED has long been a problem when it comes to code
maintainability.  Try starting small(ish) and make GEDP a non-global
variable in the MGED code.

The global variable is MGED_STATE, which is roughly equivalent to
qApp in Qt applications.  However, we try to pass it to functions
rather than just using MGED_STATE everywhere - hopefully, we'll
eventually be able to eliminate many globals and reduce the scope of
function inputs to structs specific to their uses.  Passing in the
state parameter is a first step towards that goal.  If we're using
MGED_STATE somewhere other than the initial allocation or header
definition, it's generally because some constraint of Tcl makes it
difficult to pass the state variable through the logic successfully.