Add small program to audit GED cmds against primitives list
If we want to uniformly have a GED command for each valid librt
primitive type (which makes sense, give most primitives will likely have
primitive specific controls/settings/etc.) it would be good to have a
way to detect any missing commands.  At the moment most of them don't
have commands, so this isn't something to turn on as a unit/regression
test failure, but this establishes how to do it.