Add some sanity checks for mged_state void* casts
We have to be a little careful about this thanks to the dm calls inside
doEvent, but add some ability to sanity check mged_state getting past through
voids with a magic value.

Since we don't control when doEvent gets called by Tcl/Tk, during shutdown when
we're tearing down our data structures we may hit a case where Tk wants to
handle some event after we've already closed out the mged global state
container and calls set_curr_dm.  In that particular case we need to be fault
tolerant. We're closing down and it's fine that set_curr_dm isn't going to
work.  However, we don't want to access the GEDP after free either, so check
the magic and return if things look invalid.

There is probably a better way to handle the doEvent calling of set_curr_dm to
avoid this issue completely, but trying to be minimalist about this particular
change.