Remove the global INTERP in favor of mged_state
As with GEDP, there are a few cases where passing the state through is tricky
and we fall back on using MGED_STATE, but try to pass through the pointer to
make the code less dependent on globals.

For the first pass left INTERP in the code and changed the define.