Checkpoint some work on a libbu editor selector
We've got this code in a couple places, in one form or another, and there's
nothing app or lib specific about it, so try to think about what a canonical
implementation would be that could cover all our potential use cases.  At least
one of the current implementations also encapsulates logic to launch command
line programs with xterm, but that feels like it should be a separate
consideration so leaving it out of this function.

Not hooking it up to anything yet - need to write unit tests and try on
multiple platforms first.