Integrate yedit Windows console editor into misc/tools
BRL-CAD uses text editors for a variety of commands, and has
logic to attempt selecting an editor, but the combination of Windows and
console (non-graphical) mode doesn't supply an editor out of the box on
modern systems (as of 2024 at least.)  Older Windows systems DID supply
edit.exe, so modern Windows platforms have regressed in this particular
regard.  Console mode with no option to run GUI apps is a rare
configuration for Windows, and if a user has sufficient privileges they
can install 3rd party alternatives, but the fact remains that on a
vanilla non-GUI system of that type where a user does not have either
the privileges or internet access need to install software BRL-CAD's
text editing commands will fail.

On all other targeted platforms we can usually assume at least the vi
editor (which is rather "expert friendly" but at least is present) so
Windows console mode is anomaly.  The CMD replacement shell at
https://github.com/malxau/yori has implemented a replacement for the old
edit.exe Windows console text editor, which would at least let us
supply an edit.exe-alike as a last resort.  It is MIT licensed and
should be a fairly straightforward compile on Windows.  Unfortunately
yedit is specific to Windows, so it solves the fallback problem only for
that particular case, but that seems to be about the best we can do
currently.  The Windows console environment is so different from
Unix-style environments that finding a simple, portable, liberally
licensed console text editor written in C/C++ with minimal/no
dependencies is a difficult challenge.