Rename RTG.rtg_vlfree list to rt_vlfree
Making the vlfree container a top level global since we've removed the other
RTG entries.

Thinking perhaps we should try to change the type of this to be something
hidden and specific (rather than the generic bu_list container) so we could
manage how it stores things behind the scenes, but that would have a LOT of API
implications so leaving it alone for now.