Break MGED NMG specific logic into its own file
NMG logic tends to be quite verbose, so start with that and break it out
into a separate file.  The idea here (besides making edsol easier to
digest) is to see if there are patterns that can be established for
per-primitive MGED logic that could let us set up something similar to
the OBJ functab in librt and hide per-primitive details.