Work on refactoring ARB logic
This is a bit messier than some of the others, and there is a lot more
to do here, but at least get a start.

The broader goal is to get things organized so we can do an "editing"
equivalent of the librt OBJ functab that will encapsulate all the
per-primitive editing logic.  At the moment not worrying about gloabls
being used in this code - just trying to set the stage for eliminating
all the huge primitive switch statements being used for primitive
specific editing behaviors.  Once we get that far, we can look at what
globals are in play with the editing codes and how to properly
refactor (or ideally eliminate) them.