Use functabs for both edit aware and std labels
Still using global state variables for edit-aware, but this fully clears
the ID_* switch tabel from the MGED label code.

The pattern is hinting that we'll need to be looking at an equivalent to
rt_db_internal and idb_ptr for editing - perhaps an rt_db_edit_internal
with the generic info and the per-primitive-type containers in the
idb_ptr, to correspond to how librt handles similar issues.