Fix MGED "ted" command used during solid edit
Looks like commit #447b384d4 switched editit to using ged_exec, but MGED uses
"ted" for an argv[0] here and libged doesn't know how to translate that into a
command.  editit is used in only one place, so relocate it and update it to use
the libged "editit" command name to trigger editing.