Begin pushing mged_param verification down.
mged_param has a lot of specific cases for checking if the parameters
provided as input conform to the needs of specific cases.  Need to push
this checking down to the actual methods themselves - while it's
definitely more compact to do this up front, the exposure of specific
editing methods in the PARAM_* defines means it doesn't scale well.
Not to mention we're trying to hide the details behind MGED_OBJ in the
first place.

This focuses on SSCALE and PSCALE - the latter has a second tier of
special case with es_menu flags, so metaball and pipe need to be broken
down further.  The rest of the pscale functions look as if they can get
away with one check up front, but pipe and metaball have different
requirements depending on the es_menu settings so in their cases the
checks have to happen within the local switch calls.