Try to spot and add 3 param cases
These aren't called out, and there seems to be a lot of independent
checking for params in some of the code, but try to push the full three
param checks we actually seem to need down.  Probably will be a lot of
input validation cleanup generally when we reach the point of putting
together unit tests for all this, but for now this clears one of the two
mged_param ties to edit-operation-specific code.