Don't keep running doEvent if MGED_STATE is NULL
Chris ran into a crash in MGED where doEvent was trying to access
something invalid in the course of shutting down.  We tried to
de-register the doEvent callback but that doesn't seem to be effective.
Testing a global flag did work, but Chris spotted that we could just
look for a NULL MGED_STATE instead since we do explicitly NULL that out
in mged_finish.