Move es_edflag into edit_state, and push sedit ECMD_ uses down
This one was tricky - sedit.h was using specific editing states to set
up defines which tested whether or not the current editing mode was a
rotate, translate or scale.  Rather than list out the specific states at
the top level, we instead establish four flags at the same level as
edflag which encode the same information - whether we are in a rotate,
translate, scale or pick operation.  We then define a function to handle
setting all the variables appropriately for the most common values, and
have the primitives manage them more explicitly when they are setting
one of the edflag states that would previously have activated one of the
old define definitions.  The defines themselves are updated to reference
the new high level flags rather than encoding the states directly.

Need to understand more about why so much code needs to be aware of
whether we're rotating, translating, scaling, or picking that we need
these defines, but in the interests of trying to be incremental and
limiting scope just do this to hide the primitive specific code and
defer exploring the SEDIT_* define uses until later.