Adjust the numbers being used by primitive edit modes
These are nominally private to the primitives, but it would still be
better if they don't conflict - and they can't conflict with SROT and
the other general flags.  Use the functab primitive order to prefix
the numbers, to allow us to tie any given primitive edit state flag to
its specific primitive type just from the prefix of the number itself.