move env2c and debug2c into libged.
this fixes a dependency coupling issue where misc had to come before src, but also moving because the tools are not really general.  they're highly specific to our existing repo structure and are specifically libged code generators, and only used there so looks like it should make sense to move them into place in libged.  giving it a try.