Group variables predominately used in primitives
Now that we've got the logic isolated in files, begin identifying which
variables are used mostly or exclusively in primitives code.  Those will
be the candidates for the first iteration of an rt_db_edit_internal
struct.  Conversely, anything in primitives still using edit_string
rather than s_edit from the global state will require more scrutiny.
The mousevec code seems to be updating some of the edit_absolute_*
variables, so that will take some thought.  acc_sc_sol appears to also
be in that camp, although it's currently still a global.