Move the mged_variables mv_context to mged_solid_edit
This one is a little bit more risky, since it is apparently exposed as
"context" via a structparse, but it doesn't seem (as far as I can tell)
to be playing a role in the current MGED tcl code and it is active in C
only in the primitives code.  With one exception in generic, this is the
only use of the mged_variables global in the primitives code - since
mv_context is specific to the primitives code, relocating it there.