Sync a local copy of mv_context for primitives code
Sure enough - "set context 1" is how the behavior of editing is changed
for those cases using the "mv_context" variable in the primitive editing
code.  Yikes - holy undocumented obscure unintuitive feature Batman!

For the moment, just force the MGED set command to do a sync for us so
the setting makes it into the mged_solid_edit struct and we don't have
to reference mged_variables in the primitive code.  I'll have to look at
the other uses but it might be that something like "force_keypoint"
would be a better name for that variable given how it is used in eto...