Try to update es_peqn when we might need to use it
Trying to do the update in edsol.c resulted in es_peqn's scope being
greater than just the edarb.c code - see if we can just do the
calculation when we need it in the editing code instead.  This will need
some testing to make sure we've got the equations when we need them, but
except for the edsol.c setting this variable is local to edarb.c so that
constrains the problem.

es_peqn was the last global exposed in the primitives headers, which
gets us one step closer to the primitives code being movable.