Start thinking about how to redo edit messaging
To relocate the edit code, we'll also need to get Tcl out of the
equation.  One of the prominent uses is a lot of Tcl_AppendResult calls,
so we'll need another answer there.

Probably should first look at chopping up and reworking any functions
using Tcl data types like ClientData, since they won't be able to move
as-is.