Replace Tcl_AppendResult with the log_str
This should be regarded as a first step - there is a function in the
code called "mged_print_result" which is used sometimes (but not
consistently) in the primitives I/O code.  Not completely clear yet what
it is doing, why it is there and why it's not used consistently.  It
will need to come out completely for this logic to be refactored, so the
question will ultimately be what the parent calling code needs to do.