stub initial logic for known test failures (bugs).
idea is to not add them to the main tests target, yet still use ctest definitions for them.  achieved using a external project that declares the tests, added to a custom target (bugs target in this case).

this is intended for maintaining a running backlog of known bugs that we want to fix, yet are unable to for whatever reason / prioritization. this also lets us set up the conditions and/or inputs that reproduce the bug, which might be a good proving ground for getting them fixed.  potentially a resource for new contributors to learn as well.