Remove option to use OpenMesh decimator in bot subcmd
This wasn't announced since it needed testing, fortunately. When it WAS tested
it turns out the decimation results apparently aren't always properly manifold.
For our use case that's Not Awesome, so we'll stick with the GCT version and
let OpenMesh handle the subdivision logic (which also needs be checked for
manifold preservation, but that's a separate issue...)