Restore previous src/hdrs add_subdirectory order to fix BRLCAD_COMPONENTS
The BRLCAD_COMPONENTS logic for building only subsets of BRL-CAD uses the
target definitions in src to enable/disable install rules for per-library
headers.  We need the lib targets defined before we define our header install
rules, or we don't know which headers are needed and almost none of them get
installed.