Test having s_edit actually be transient
Up until now all we've been doing is moving globals into structs and
renaming, but this makes s_edit something that is created and destroyed
when MGED enters and exits the sed and oed editing states.

This will need a ton of testing - basically, any point that's going to
edit an object in earnest will need to do the setup and tear-down for
the mged_solid_edit state structure.  The "sed" and "oed" commands are
the obvious entry points, and accept/reject the obvious exits, but there
may be more than just those.

This is one of the key changes that will be necessary for this logic to
move to a library - we have to properly divorce ourselves from the
persistent MGED global state.  Everything in primitives (except for the
code still using Tcl directly) needs to try to move to taking a solid
edit struct rather than mged_state itself.