Stub in librt functab entries for primitive matrix application
Right now, to apply a matrix to an internal representation of a solid we're
going through the export5/import5 cycle - the import step is where the matrix
gets applied.  That's fine if we're doing I/O, but for edit operations using
that API means we need the dbip just to apply the matrix.  Most of the time
that shouldn't be needed - rt_db_internals (with a few execptions like extrude)
should be able to function independent of their parent database.

Will need to be done carefully, but set up to explore if we can make the matrix
application logic in the *import5 routines available as a direct stand-alone
functab call.