Get MGED compiling using clbks for nmg and extrude
Extrude in particular we will probably want to handle with an internal struct
rather than a callback, but until we get the globals folded into editing
internal structs we don't have a dbip for the lookup.

Not ready for merging to main, but this removes explicit uses of Tcl* functions
and mged_state from the primitives code.