Use the define rather than a raw number
Arguably this isn't strictly accurate, since libnmg I/O versioning isn't
TECHNICALLY tied in a data sense to .g I/O versions, but functionally it should
be fine and provides at least a little context.

The broader problem here is needing to use versioned librt routines in an API
that only really wants to be working with rt_db_internal structs.  Wondering if
it wouldn't be cleaner to encode versioning of the idb_ptr data in the
rt_db_internal container rather than needing to propagate the dbip for that...