Migrate some of the edit logic to librt
Having to stub in a couple things to make this incremental, but start working
on moving the functional edit logic to librt proper.  We want unit tests for
this that don't depend on MGED, so before defining them we want the logic in
the lib context rather than the app context.