Back MGED up to commit 3998b4e108
Quite a bit more work to be done to get editing working and tested with
the refactor, so back things up to before the code moves began.

Leaving the librt editing code in place, so we can merge back into main
and then branch again to continue working with MGED and the librt code.
This state of MGED doesn't use any of the librt editing APIs, so they'll
be dormant until we get things properly retuned and tested.