Restore (in a branch) the in-progress MGED using librt editing
OK, main has a version of MGED from before the code relocations started.
We now make another branch to focus on shaking down MGED's editing
capabilities with this new form of the edit code.

Doing it this way so we (hopefully) wind up with fewer conflicts when
rebasing and/or merging main - prior branch was getting a bit unwieldy.

This reverts commit 187b138fd2f884392305491652f394d04281febc.