Remove edit_menu in favor of edit_flag
In the case of arb8, which has more complex/extensive use, make it an
entry in the arb specific struct - otherwise, just use edit_flag.  In
the case of PSCALE, this means rather than using a LIBRT level PSCALE
define and edit_menu to get specific, the primitives take care of things
on an individual basis.  The top level variable becomes only convenience
for updating the flags.