Adjust edit_xy methods, add sedit equiv. call
Ah, OK.  The edit_xy routines merely set the stage for the full-on
processing edit call.  Initial breakout was slightly off the mark in
that it was having the edit_xy calls try to do everything in certain
cases and not calling the main routine from sedit_mouse, which is why
most things weren't updating properly.

Will need another pass to make sure all the equivalents are in place as
compared to the big case table, but with this change setting the A
parameter in cone.s in moss.g example is behaving much better.